dev/tests/requirements.txt
bcrypt==4.0.1
psycopg2-binary==2.9.5
PyYAML==6.0
requests==2.28.1
robotframework==6.0.1
robotframework-databaselibrary==1.2.4
robotframework-pythonlibcore==4.0.0
robotframework-requests==0.9.4
robotframework-seleniumlibrary==6.0.0
robotframework-sshlibrary==3.8.0
robotframework-pabot==2.10.0
rpmfile==1.0.8
selenium==4.7.2
